requirements.txt
Flask==2.3.3
ultralytics==8.0.196
Pillow==10.0.1
numpy==1.24.3
pytesseract==0.3.10
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2 